+++ packages/python/plotly/plotly/graph_objs/_figure.py
@@ -3980,7 +3980,9 @@ class Figure(BaseFigure):
         showscale=None,
         stream=None,
         text=None,
+        textfont=None,
         textsrc=None,
+        texttemplate=None,
         transpose=None,
         uid=None,
         uirevision=None,
@@ -4207,9 +4209,29 @@ class Figure(BaseFigure):
             or dict with compatible properties
         text
             Sets the text elements associated with each z value.
+        textfont
+            For this trace it only has an effect if `coloring` is
+            set to "heatmap". Sets the text font.
         textsrc
             Sets the source reference on Chart Studio Cloud for
             `text`.
+        texttemplate
+            For this trace it only has an effect if `coloring` is
+            set to "heatmap". Template string used for rendering
+            the information text that appear on points. Note that
+            this will override `textinfo`. Variables are inserted
+            using %{variable}, for example "y: %{y}". Numbers are
+            formatted using d3-format's syntax
+            %{variable:d3-format}, for example "Price: %{y:$.2f}".
+            https://github.com/d3/d3-format/tree/v1.4.5#d3-format
+            for details on the formatting syntax. Dates are
+            formatted using d3-time-format's syntax
+            %{variable|d3-time-format}, for example "Day:
+            %{2019-01-01|%A}". https://github.com/d3/d3-time-
+            format/tree/v2.2.3#locale_format for details on the
+            date formatting syntax. Every attributes that can be
+            specified per-point (the ones that are `arrayOk: true`)
+            are available. variables `x`, `y`, `z` and `text`.
         transpose
             Transposes the z data.
         uid
@@ -4434,7 +4456,9 @@ class Figure(BaseFigure):
             showscale=showscale,
             stream=stream,
             text=text,
+            textfont=textfont,
             textsrc=textsrc,
+            texttemplate=texttemplate,
             transpose=transpose,
             uid=uid,
             uirevision=uirevision,
@@ -6098,7 +6122,9 @@ class Figure(BaseFigure):
         showscale=None,
         stream=None,
         text=None,
+        textfont=None,
         textsrc=None,
+        texttemplate=None,
         transpose=None,
         uid=None,
         uirevision=None,
@@ -6316,9 +6342,27 @@ class Figure(BaseFigure):
             or dict with compatible properties
         text
             Sets the text elements associated with each z value.
+        textfont
+            Sets the text font.
         textsrc
             Sets the source reference on Chart Studio Cloud for
             `text`.
+        texttemplate
+            Template string used for rendering the information text
+            that appear on points. Note that this will override
+            `textinfo`. Variables are inserted using %{variable},
+            for example "y: %{y}". Numbers are formatted using
+            d3-format's syntax %{variable:d3-format}, for example
+            "Price: %{y:$.2f}".
+            https://github.com/d3/d3-format/tree/v1.4.5#d3-format
+            for details on the formatting syntax. Dates are
+            formatted using d3-time-format's syntax
+            %{variable|d3-time-format}, for example "Day:
+            %{2019-01-01|%A}". https://github.com/d3/d3-time-
+            format/tree/v2.2.3#locale_format for details on the
+            date formatting syntax. Every attributes that can be
+            specified per-point (the ones that are `arrayOk: true`)
+            are available. variables `x`, `y`, `z` and `text`.
         transpose
             Transposes the z data.
         uid
@@ -6544,7 +6588,9 @@ class Figure(BaseFigure):
             showscale=showscale,
             stream=stream,
             text=text,
+            textfont=textfont,
             textsrc=textsrc,
+            texttemplate=texttemplate,
             transpose=transpose,
             uid=uid,
             uirevision=uirevision,
@@ -6929,6 +6975,8 @@ class Figure(BaseFigure):
         autobinx=None,
         autobiny=None,
         bingroup=None,
+        cliponaxis=None,
+        constraintext=None,
         cumulative=None,
         customdata=None,
         customdatasrc=None,
@@ -6945,6 +6993,8 @@ class Figure(BaseFigure):
         hovertextsrc=None,
         ids=None,
         idssrc=None,
+        insidetextanchor=None,
+        insidetextfont=None,
         legendgroup=None,
         legendgrouptitle=None,
         legendrank=None,
@@ -6957,12 +7007,17 @@ class Figure(BaseFigure):
         offsetgroup=None,
         opacity=None,
         orientation=None,
+        outsidetextfont=None,
         selected=None,
         selectedpoints=None,
         showlegend=None,
         stream=None,
         text=None,
+        textangle=None,
+        textfont=None,
+        textposition=None,
         textsrc=None,
+        texttemplate=None,
         uid=None,
         uirevision=None,
         unselected=None,
@@ -7021,6 +7076,14 @@ class Figure(BaseFigure):
             "overlay" and on different axes (of the same axis type)
             can have compatible bin settings. Note that histogram
             and histogram2d* trace can share the same `bingroup`
+        cliponaxis
+            Determines whether the text nodes are clipped about the
+            subplot axes. To show the text nodes above axis lines
+            and tick labels, make sure to set `xaxis.layer` and
+            `yaxis.layer` to *below traces*.
+        constraintext
+            Constrain the size of text inside or outside a bar to
+            be no larger than the bar itself.
         cumulative
             :class:`plotly.graph_objects.histogram.Cumulative`
             instance or dict with compatible properties
@@ -7118,6 +7181,11 @@ class Figure(BaseFigure):
         idssrc
             Sets the source reference on Chart Studio Cloud for
             `ids`.
+        insidetextanchor
+            Determines if texts are kept at center or start/end
+            points in `textposition` "inside" mode.
+        insidetextfont
+            Sets the font used for `text` lying inside the bar.
         legendgroup
             Sets the legend group for this trace. Traces part of
             the same legend group hide/show at the same time when
@@ -7177,6 +7245,8 @@ class Figure(BaseFigure):
             Sets the orientation of the bars. With "v" ("h"), the
             value of the each bar spans along the vertical
             (horizontal).
+        outsidetextfont
+            Sets the font used for `text` lying outside the bar.
         selected
             :class:`plotly.graph_objects.histogram.Selected`
             instance or dict with compatible properties
@@ -7198,9 +7268,43 @@ class Figure(BaseFigure):
             single string, the same string appears over all bars.
             If an array of string, the items are mapped in order to
             the this trace's coordinates.
+        textangle
+            Sets the angle of the tick labels with respect to the
+            bar. For example, a `tickangle` of -90 draws the tick
+            labels vertically. With "auto" the texts may
+            automatically be rotated to fit with the maximum size
+            in bars.
+        textfont
+            Sets the text font.
+        textposition
+            Specifies the location of the `text`. "inside"
+            positions `text` inside, next to the bar end (rotated
+            and scaled if needed). "outside" positions `text`
+            outside, next to the bar end (scaled if needed), unless
+            there is another bar stacked on this one, then the text
+            gets pushed inside. "auto" tries to position `text`
+            inside the bar, but if the bar is too small and no bar
+            is stacked on this one the text is moved outside. If
+            "none", no text appears.
         textsrc
             Sets the source reference on Chart Studio Cloud for
             `text`.
+        texttemplate
+            Template string used for rendering the information text
+            that appear on points. Note that this will override
+            `textinfo`. Variables are inserted using %{variable},
+            for example "y: %{y}". Numbers are formatted using
+            d3-format's syntax %{variable:d3-format}, for example
+            "Price: %{y:$.2f}".
+            https://github.com/d3/d3-format/tree/v1.4.5#d3-format
+            for details on the formatting syntax. Dates are
+            formatted using d3-time-format's syntax
+            %{variable|d3-time-format}, for example "Day:
+            %{2019-01-01|%A}". https://github.com/d3/d3-time-
+            format/tree/v2.2.3#locale_format for details on the
+            date formatting syntax. Every attributes that can be
+            specified per-point (the ones that are `arrayOk: true`)
+            are available. variables `label` and `value`.
         uid
             Assign an id to this trace, Use this to provide object
             constancy between traces during animations and
@@ -7319,6 +7423,8 @@ class Figure(BaseFigure):
             autobinx=autobinx,
             autobiny=autobiny,
             bingroup=bingroup,
+            cliponaxis=cliponaxis,
+            constraintext=constraintext,
             cumulative=cumulative,
             customdata=customdata,
             customdatasrc=customdatasrc,
@@ -7335,6 +7441,8 @@ class Figure(BaseFigure):
             hovertextsrc=hovertextsrc,
             ids=ids,
             idssrc=idssrc,
+            insidetextanchor=insidetextanchor,
+            insidetextfont=insidetextfont,
             legendgroup=legendgroup,
             legendgrouptitle=legendgrouptitle,
             legendrank=legendrank,
@@ -7347,12 +7455,17 @@ class Figure(BaseFigure):
             offsetgroup=offsetgroup,
             opacity=opacity,
             orientation=orientation,
+            outsidetextfont=outsidetextfont,
             selected=selected,
             selectedpoints=selectedpoints,
             showlegend=showlegend,
             stream=stream,
             text=text,
+            textangle=textangle,
+            textfont=textfont,
+            textposition=textposition,
             textsrc=textsrc,
+            texttemplate=texttemplate,
             uid=uid,
             uirevision=uirevision,
             unselected=unselected,
@@ -7407,6 +7520,8 @@ class Figure(BaseFigure):
         showlegend=None,
         showscale=None,
         stream=None,
+        textfont=None,
+        texttemplate=None,
         uid=None,
         uirevision=None,
         visible=None,
@@ -7644,6 +7759,24 @@ class Figure(BaseFigure):
         stream
             :class:`plotly.graph_objects.histogram2d.Stream`
             instance or dict with compatible properties
+        textfont
+            Sets the text font.
+        texttemplate
+            Template string used for rendering the information text
+            that appear on points. Note that this will override
+            `textinfo`. Variables are inserted using %{variable},
+            for example "y: %{y}". Numbers are formatted using
+            d3-format's syntax %{variable:d3-format}, for example
+            "Price: %{y:$.2f}".
+            https://github.com/d3/d3-format/tree/v1.4.5#d3-format
+            for details on the formatting syntax. Dates are
+            formatted using d3-time-format's syntax
+            %{variable|d3-time-format}, for example "Day:
+            %{2019-01-01|%A}". https://github.com/d3/d3-time-
+            format/tree/v2.2.3#locale_format for details on the
+            date formatting syntax. Every attributes that can be
+            specified per-point (the ones that are `arrayOk: true`)
+            are available. variable `z`
         uid
             Assign an id to this trace, Use this to provide object
             constancy between traces during animations and
@@ -7836,6 +7969,8 @@ class Figure(BaseFigure):
             showlegend=showlegend,
             showscale=showscale,
             stream=stream,
+            textfont=textfont,
+            texttemplate=texttemplate,
             uid=uid,
             uirevision=uirevision,
             visible=visible,
@@ -7905,6 +8040,8 @@ class Figure(BaseFigure):
         showlegend=None,
         showscale=None,
         stream=None,
+        textfont=None,
+        texttemplate=None,
         uid=None,
         uirevision=None,
         visible=None,
@@ -8156,6 +8293,26 @@ class Figure(BaseFigure):
         stream
             :class:`plotly.graph_objects.histogram2dcontour.Stream`
             instance or dict with compatible properties
+        textfont
+            For this trace it only has an effect if `coloring` is
+            set to "heatmap". Sets the text font.
+        texttemplate
+            For this trace it only has an effect if `coloring` is
+            set to "heatmap". Template string used for rendering
+            the information text that appear on points. Note that
+            this will override `textinfo`. Variables are inserted
+            using %{variable}, for example "y: %{y}". Numbers are
+            formatted using d3-format's syntax
+            %{variable:d3-format}, for example "Price: %{y:$.2f}".
+            https://github.com/d3/d3-format/tree/v1.4.5#d3-format
+            for details on the formatting syntax. Dates are
+            formatted using d3-time-format's syntax
+            %{variable|d3-time-format}, for example "Day:
+            %{2019-01-01|%A}". https://github.com/d3/d3-time-
+            format/tree/v2.2.3#locale_format for details on the
+            date formatting syntax. Every attributes that can be
+            specified per-point (the ones that are `arrayOk: true`)
+            are available. variables `x`, `y`, `z` and `text`.
         uid
             Assign an id to this trace, Use this to provide object
             constancy between traces during animations and
@@ -8346,6 +8503,8 @@ class Figure(BaseFigure):
             showlegend=showlegend,
             showscale=showscale,
             stream=stream,
+            textfont=textfont,
+            texttemplate=texttemplate,
             uid=uid,
             uirevision=uirevision,
             visible=visible,